wikipedia-style traversal: cross-file parser, focus/trail, dismiss cascade
- parser walks local relative imports across files; split into
  parse_js.mjs / visitors.mjs / resolve.mjs so the demo has real
  cross-file edges
- focus/trail navigation layered on top of the nested file-card view —
  click a row to focus (sticky-expand its ancestors), click again to
  toggle expansion, click a trail row to jump back
- cards accumulate as you navigate; each new card stamps :opened-by so
  dismiss cascades to descendants
- column layout anchored to opener (right of, stacked below siblings)
- camera tweens to fit-bbox after every layout change
- edges anchor at call-site line via offsetStart when source is rendered
