+++ public/index.html
@@ -106,6 +106,15 @@
     }
     .file-card .row.expanded { background: #1c1c25; }
     .file-card .row.expanded + .row { border-top-color: #0c0c12; }
+    /* Focused row: current "you are here" in the Wikipedia-style walk.
+       Trail rows: breadcrumb of previously-focused rows. Both act as
+       subtle color cues layered on top of the normal expanded treatment. */
+    .file-card .row.focused > .row-head { box-shadow: inset 3px 0 0 #ffaa55; }
+    .file-card .row.focused > .row-head .name { color: #ffaa55; }
+    .file-card .row.expanded.focused > .row-head .name { color: #ffc88a; }
+    .file-card .row.trail > .row-head { box-shadow: inset 3px 0 0 #5a8aff; }
+    .file-card .row.trail > .row-head .name { color: #8fa8ff; }
+    .file-card .row.expanded.trail > .row-head .name { color: #aabfff; }
     /* Direct-child selectors so nested rows don't inherit the
        dark-background color treatment of their expanded parent. */
     .file-card .row.expanded > .row-head .kind { color: #8c8ca0; }
@@ -156,6 +165,26 @@
       font-family: ui-monospace, SFMono-Regular, monospace;
       font-size: 11px;
       cursor: grab;
+      display: flex;
+      align-items: center;
+      gap: 6px;
+    }
+    .file-card .path .dismiss {
+      display: inline-flex;
+      align-items: center;
+      justify-content: center;
+      width: 16px;
+      height: 16px;
+      border-radius: 50%;
+      background: rgba(255, 255, 255, 0.06);
+      color: #8c8c9a;
+      cursor: pointer;
+      font-size: 13px;
+      line-height: 1;
+    }
+    .file-card .path .dismiss:hover {
+      background: #d44;
+      color: #fff;
     }
   </style>
 </head>
